Remove Mention of Terminal User Inteface (#7)
Progression towards an Electron app, in order to better display
information. First goal is to create a Rust CLI app that can give the
sought after information and then wrap that into a wasm-binding for the
performance improvements. Then use Electron + Deno + React to build an
actual interface. Production on the CLI app is first and foremost,
focusing on just displaying information for the project as a whole.