Improve library usability (#9)
Moved from `GitReference` `Enum` to individual structs and `Trait` in order to bring the same functionality for `Repo::checkout`. 
this is more usable and exposes an API that's more like the standard library. 

Instead of providing name and name_bytes, it attempts to
convert using `String::from_ut8`. Also Uses `Chrono` for times in order to
be better user facing.